refactor: update TrustGauge to consume tiers.ts
